Title: a/b 출력, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240924/a/b 출력/a-divide-b.cpp
@@ -16,7 +16,7 @@ int main() {
     {
         c *= 10.f;
         num = c;
-        c -= num;
+        c -= (float)num;
         cout << num % 10;
     }
     return 0;
